backend/order-gateway/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pyjwt>=2.8.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
redis>=5.0.3
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
anyio[trio]>=4.3.0
